FinSight/backend/requirements.txt
absl-py==2.1.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
azure-core==1.32.0
azure-identity==1.21.0
Babel==2.14.0
beautifulsoup4==4.12.2
black==23.1.0
bleach==6.1.0
blessed==1.20.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.2
category-encoders==2.6.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.2.0
click==8.0.4
cloudpickle==3.0.0
cmake==3.31.6
comm==0.2.2
contourpy==1.2.0
copyedit==1.0.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.10
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dlib==19.24.6
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.19.1
filelock==3.12.4
flake8==5.0.4
flake8-nb==0.5.2
Flask==3.0.3
flatbuffers==25.1.24
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.6.0
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.37
google-ai-generativelanguage==0.6.15
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-secret-manager==2.23.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.62.0
googledrivedownloader==0.4
grpc-google-iam-v1==0.14.2
grpcio==1.70.0
grpcio-status==1.63.0rc1
h11==0.14.0
h5py==3.12.1
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.4
imbalanced-learn==0.12.3
importlib-metadata==6.8.0
inquirer==3.1.3
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.0.4
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==3.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.11
kaggle==1.5.12
kaleido==0.2.1
keras==3.8.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightgbm==4.3.0
litellm==0.1.738
llvmlite==0.42.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.0.4
music21==8.1.0
mutagen==1.47.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nodeenv==1.8.0
notebook_shim==0.2.3
numba==0.59.1
numpy==1.26.3
oauthlib==3.2.2
open-interpreter==0.1.4
openai==0.27.10
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.4
overrides==7.4.0
packaging==24.2
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
plotly==5.22.0
plotly-resampler==0.10.0
pmdarima==2.0.4
pooch==1.8.2
prometheus-client==0.19.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==4.25.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-sudoku==1.0.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycaret==3.3.2
pycparser==2.21
pycryptodomex==3.19.1
pydantic==2.10.6
pydantic_core==2.27.2
pydot==1.4.2
pygame==2.5.2
Pygments==2.19.1
PyJWT==2.10.1
pyod==2.0.0
pyparsing==3.1.1
pyperclip==1.9.0
pyright==1.1.354
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
python-louvain==0.16
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.1
rdflib==7.0.0
readchar==4.0.5
referencing==0.32.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rpds-py==0.16.2
rsa==4.9
schemdraw==0.15
scikit-base==0.7.8
scikit-image==0.19.3
scikit-learn==1.2.1
scikit-plot==0.3.7
scipy==1.10.0
semantic-version==2.10.0
Send2Trash==1.8.2
setuptools-rust==1.9.0
six==1.17.0
sktime==0.26.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.13.1
soupsieve==2.5
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.2
sudokum==0.0.1
sympy==1.12.1
tbats==1.1.3
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-addons==0.23.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
tensorflow-probability==0.18.0
termcolor==2.5.0
terminado==0.18.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.14.0
tokentrim==0.1.10
torch==2.3.0
torch_geometric==1.7.2
torchvision==0.18.0
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.3
tsdownsample==0.1.3
txt==2020.11.12
typeguard==2.13.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.34.0
watchdog==3.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.3
wget==3.2
widgetsnbextension==4.0.11
wrapt==1.17.2
wurlitzer==3.1.0
xarray==2024.5.0
xxhash==3.4.1
yarl==1.9.2
yellowbrick==1.5
youtube-dl==2021.12.17
yt-dlp==2023.12.30
zipp==3.17.0
